requirements.txt
pandas
numpy
matplotlib
Pillow
requests
python-dotenv
entsoe-py